clear cldnn unit test issue (#75)
* [GPU][Dynamic] Fix cldnn unit tests for gemm and depth_to_space
- Modify test params in gemm gpu test
- Modify test params in depth_to_space gpu tests
- Modify test params in strided_splice gpu tests
- Add Simple PartialShape test

* [GPU][Dynamic] Fix premute gpu test issue
- Modify check condition in layout_optimizer::can_fuse_reorder_to_prev

* [GPU][Dynamic] Fix data_layout_test.size_check issues
- implement layout::get_ordered_dims

* [GPU][Dynamic] Fix eltwsie fusing in fusings_gpu issues
- Change get_dims() to get_tensor().sizes()

* [GPU][Dynamic] Fix test params
- experimental_detectron_roi_feature_extractor_gpu_fp32.multiple_feature_extractor_op_with_different_number_of_inputs
- reorder_gpu_f32.bfwzyx_bfyx_chain

* [GPU][Dynamic] Fix reshape test issues
- Remove checking dynamic with 2 input for reshape
- Add wait_for_event to get real data after input node execution
- Modify test params for reshape test

* [GPU][Dynamic] Fix reduce test issues
- Fix output layout calculation in reduce_ist::calc_output_layout

* [GPU][Dynamic] Fix gather_nd and gatter issue
- Fix batch_dim issue which always convert negative value to positive without calcuation
- Fix test expected output result
- Ajust output shape by output format
- Add gather unit test for dynamic shape

* [GPU][Dynamic] Fix fusings_gpu issues
- Fix permute order
- Fix test params in gather, loop, and permute fustion test
- Fix reference function and test params in one_hot_gpu_test
- Skip one_hot_error.basic_error_bad_shape test
- Fix cache_test multi-threading issue
- Uncomment fully connected kernel selectors

* [GPU][Dynamic] Fix scatter update issues
- Change type of axis from int to size_t
- Fix axis in fusion test and single test

+++ src/plugins/intel_gpu/src/graph/impls/ocl/scatter_update.cpp
@@ -14,7 +14,7 @@ using namespace cldnn;
 
 namespace cldnn {
 namespace ocl {
-kernel_selector::scatter_update_axis convert_axis(const int axis, const scatter_update_node& arg) {
+kernel_selector::scatter_update_axis convert_axis(const size_t axis, const scatter_update_node& arg) {
     auto rank = arg.input(0).get_output_layout().get_rank();
     auto cldnn_axis = axis;
     if (axis >= 2) {
